requirements.txt
# Token Listing FDV Benchmark Tool
# Python >= 3.10 required

# Web dashboard
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0

# Core dependencies
pydantic>=2.0.0
httpx>=0.24.0
pyyaml>=6.0

# Exchange data
ccxt>=4.0.0

# CLI reporting
typer>=0.9.0
rich>=13.0.0
